movie/requirements.txt
requests
Pillow
tk
bs4